Backend/requirements.txt
flask==2.2.2
flask_cors==3.0.10
werkzeug==2.3.8
Pillow==10.0
pytorch==2.1.0 
torchvision==0.16.0 
torchaudio==2.1.0
openai==1.6.0
networkx==3.1
matplotlib==3.8.2